moved daily built files out of active use, updated router
