+++ src/app/app-routing.module.ts
@@ -2,45 +2,47 @@ import { NgModule } from '@angular/core';
 import { RouterModule, Routes,  TitleStrategy, RouterStateSnapshot } from '@angular/router';
 // https://www.telerik.com/blogs/angular-basics-router-link-overview
 import { GuestProfileComponent } from './built/guest-profile/guest-profile.component';
-import { SitemapComponent } from './sitemap/sitemap.component';
 import { HeaderComponent } from './header/header.component';
 import { PageNotFoundComponent } from './page-not-found/page-not-found.component';
+import { PageErrorComponent } from './page-error/page-error.component';
 // import { A01c2GuestAboutComponent } from './built/a01c2-guest-about/a01c2-guest-about.component'
 // import { A01c4GuestWebLinksComponent } from './built/a01c4-guest-web-links/a01c4-guest-web-links.component'
 // import { A02aUpdateYourPasswordComponent } from './built/a02a-update-your-password/a02a-update-your-password.component'
-
-import { A01aGuestSignInComponent } from './a01a-guest-sign-in/a01a-guest-sign-in.component';
-import { A01aGuestSignUpComponent } from './a01a-guest-sign-up/a01a-guest-sign-up.component';
+// import { A01aGuestSignInComponent } from './built/a01a-guest-sign-in/a01a-guest-sign-in.component';
+// import { A01aGuestSignUpComponent } from './built/a01a-guest-sign-up/a01a-guest-sign-up.component';
 // import { A04aHamburgerNavComponent } from './built/a04a-hamburger-nav/a04a-hamburger-nav.component';
 // import { A05bGuestPublicLandingComponent } from './built/a05b-guest-public-landing/a05b-guest-public-landing.component';
 // import { B04aVendorAboutLandingDisplayComponent } from './built/b04a-vendor-about-landing-display/b04a-vendor-about-landing-display.component';
 import { FooterComponent } from './footer/footer.component';
+import { B04aVendorAboutLandingDisplayComponent } from './built/b04a-vendor-about-landing-display/b04a-vendor-about-landing-display.component';
+// import { A01bGuestWelcomeComponent } from './built/a01b-guest-welcome/a01b-guest-welcome.component';
+// import { A01c3GuestSocialMediaComponent } from './built/a01c3-guest-social-media/a01c3-guest-social-media.component';
+import { A01bGuestWelcomeBackComponent } from './a01b-guest-welcome-back/a01b-guest-welcome-back.component';
 
 
-import { B04aVendorAboutLandingDisplayComponent } from './b04a-vendor-about-landing-display/b04a-vendor-about-landing-display.component';
-import { A01bGuestWelcomeComponent } from './a01b-guest-welcome/a01b-guest-welcome.component';
-import { A01c3GuestSocialMediaComponent } from './a01c3-guest-social-media/a01c3-guest-social-media.component';
+
 
 const routes: Routes = [
   { path: 'guest', component: GuestProfileComponent },
   // { path: 'guest-about', component: A01c2GuestAboutComponent },
   // { path: 'guest-landing', component: A05bGuestPublicLandingComponent},
+  // { path: 'guest-social', component: A01c3GuestSocialMediaComponent},
   // { path: 'guest-weblinks', component: A01c4GuestWebLinksComponent },
-  { path: 'guest-sign-in', component: A01aGuestSignInComponent},
-  { path: 'guest-sign-up', component: A01aGuestSignUpComponent},
-  
+  // { path: 'guest-sign-in', component: A01aGuestSignInComponent},
+  // { path: 'guest-sign-up', component: A01aGuestSignUpComponent},
   // { path: 'update-password', component: A02aUpdateYourPasswordComponent},
   // { path: 'vendor-landing', component: B04aVendorAboutLandingDisplayComponent},
   // { path: 'hamburger-nav', component: A04aHamburgerNavComponent},
-  { path: 'sitemap', component: SitemapComponent },
   { path: 'header', component: HeaderComponent },
   { path: 'footer', component: FooterComponent },
   { path: 'page-not-found', component: PageNotFoundComponent },
-  { path: 'guest-welcome', component: A01bGuestWelcomeComponent},
-  { path: 'update-password', component: A02aUpdateYourPasswordComponent},
-  { path: 'hamburger-nav', component: A04aHamburgerNavComponent},
-  { path: 'guest-landing', component: A05bGuestPublicLandingComponent},
-  { path: 'guest-social', component: A01c3GuestSocialMediaComponent},
+  { path: 'page-error', component: PageErrorComponent },
+  // { path: 'guest-welcome', component: A01bGuestWelcomeComponent},
+  { path: 'guest-welcome-back', component: A01bGuestWelcomeBackComponent},
+  // { path: 'update-password', component: A02aUpdateYourPasswordComponent},
+  // { path: 'hamburger-nav', component: A04aHamburgerNavComponent},
+  // { path: 'guest-landing', component: A05bGuestPublicLandingComponent},
+  
   { path: 'vendor-landing', component: B04aVendorAboutLandingDisplayComponent},
   { path: '',  redirectTo: 'guest', pathMatch: 'full'},
   { path: "**", redirectTo: 'page-not-found' },
