--- src/app/sitemap/sitemap.component.spec.ts
@@ -1,23 +1,0 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { SitemapComponent } from './sitemap.component';
-
-describe('SitemapComponent', () => {
-  let component: SitemapComponent;
-  let fixture: ComponentFixture<SitemapComponent>;
-
-  beforeEach(async () => {
-    await TestBed.configureTestingModule({
-      declarations: [ SitemapComponent ]
-    })
-    .compileComponents();
-
-    fixture = TestBed.createComponent(SitemapComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
